+++ src/modules/routine/application/use-cases/create-routine.ts
@@ -7,7 +7,7 @@ import {
 } from '@/modules/routine/domain/errors/routine.errors';
 import { CYCLE_TYPES } from '@/modules/routine/domain/constants/routine.constants.cycle-types';
 import type { CreateRoutineInput } from '@/modules/routine/application/dtos/create-routine.dto';
-import type { IdGeneratorRepository } from '@/shared/application/id-generator';
+import type { IdGeneratorRepository } from '@/shared/application/id-generator.repository';
 import type { IRoutineRepository } from '@/modules/routine/domain/routine.repository';
 import type { ISessionRepository } from '@/modules/session/domain/session.repository';
 import type { RoutineProps } from '@/modules/routine/domain/routine.types';
